split hits script intocomponent parts hit/alert/heading
