+++ app/frontend/src/search/ui/heading.test.js
@@ -1,4 +1,4 @@
-import { createHeadingMarkup, getSearchTermsPrefix, createHeadingHTMLForSearchTerm, getJobAlertLink, getJobAlertLinkParam } from './hits';
+import { createHeadingMarkup, getSearchTermsPrefix, createHeadingHTMLForSearchTerm } from './heading';
 
 describe('createHeadingMarkup', () => {
     test('creates formatted and readable markup for number of results with no search terms', () => {
@@ -51,18 +51,5 @@ describe('getSearchTermsPrefix', () => {
 describe('createHeadingHTMLForSearchTerm', () => {
     test('returns a string of capitalized words after a prefix', () => {
         expect(createHeadingHTMLForSearchTerm('176 jobs near', 'babylon gardens', true)).toBe('176 jobs near <span class="govuk-!-font-weight-bold text-capitalize text-wrap-apostrophe">babylon gardens</span>');
-    });
-});
-
-describe('getJobAlertLink', () => {
-    test('returns a string of capitalized words after a prefix', () => {
-        expect(getJobAlertLink('/jobs?utf8=✓&keyword=&location=south%20gloucestershire')).toBe('/subscriptions/new?search_criteria%5Blocation%5D=south+gloucestershire&search_criteria%5Blocation_category%5D=south+gloucestershire&');
-        expect(getJobAlertLink('/jobs?utf8=✓&keyword=maths&location=south%20gloucestershire&radius=10')).toBe('/subscriptions/new?search_criteria%5Bkeyword%5D=maths&search_criteria%5Blocation%5D=south+gloucestershire&search_criteria%5Blocation_category%5D=south+gloucestershire&search_criteria%5Bradius%5D=10&');
-    });
-});
-
-describe('getJobAlertLinkParam', () => {
-    test('converts key,value to valid url parameter string', () => {
-        expect(getJobAlertLinkParam('location', 'south east')).toBe('search_criteria%5Blocation%5D=south+east&');
     });
 });
