+++ app/frontend/src/search/ui/hits.test.js
@@ -1,0 +1,5 @@
+describe('createHeadingHTMLForSearchTerm', () => {
+    test('returns a string of capitalized words after a prefix', () => {
+        expect(1).toBe(1);
+    });
+});
